+++ common/src/main/java/com/aws/jverify/common/TestMarkup.java
@@ -288,3 +288,13 @@ class TextBuffer {
 }
 
 record Span(Integer start, String content) {}
+
+class AnnotatedSpan {
+    public final String annotation;
+    public final TextSpan span;
+
+    public AnnotatedSpan(String annotation, TextSpan span) {
+        this.annotation = annotation;
+        this.span = span;
+    }
+}
